pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kvdeveloper"
dynamic = ["version"]
dependencies = [
  "kivy>=2.0.0",
  "pillow>=10.0.0",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "requests>=2.32.0",
  "qrcode",
]
requires-python = ">=3.9"
authors = [
  {name = "Kartavya Shukla", email = "novfensec@protonmail.com"},
]
maintainers = [
  {name = "Kartavya Shukla", email = "novfensec@protonmail.com"},
]
description = "KvDeveloper is a PyPI module designed to streamline the development of Kivy and KivyMD applications. Inspired by Expo CLI for React Native, KvDeveloper provides starter templates and essential functionalities to kickstart your projects with ease."
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["kivy", "kivymd", "open source", "kvdeveloper", "kartavya shukla", "novfensec", "kv"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Build Tools",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3"
]

[project.urls]
Homepage = "https://github.com/Novfensec/KvDeveloper"
Documentation = "https://novfensec.github.io/KvDeveloper.docs"
Repository = "https://github.com/Novfensec/KvDeveloper"
"Bug Tracker" = "https://github.com/Novfensec/KvDeveloper/issues"
Changelog = "https://github.com/Novfensec/KvDeveloper/blob/main/CHANGELOG.md"

[tool.setuptools.dynamic]
version = {attr = "kvdeveloper.__version__"}

[project.scripts]
kvdeveloper = "kvdeveloper.__main__:main"
kvd = "kvdeveloper.__main__:main"

[project.optional-dependencies]
all = [
  "markdown2",
  "pyqt5",
  "pyqtwebengine",
]

[tool.setuptools]
package-dir = {"kvdeveloper" = "kvdeveloper"}

[tool.setuptools.package-data]
"kvdeveloper" = [
  "**/*.md",
  "**/*.txt",
  "**/*.ipynb",
  "**/*.kv",
  "**/*.yml",
  "**/*.png",
  "**/*.spec",
  "**/*.js",
  "**/*.css",
  "**/*.toml",
  "templates/p4a/build.tmpl.gradle",
  "templates/p4a/gradle.json",
]

[tool.setuptools.exclude-package-data]
"kvdeveloper" = ["*.pyc", "*.pyo"]
